Изменен компонент Catalog и прокси, осуществлен рефакторинг utils
