+++ src/components/main-screen/main-screen.jsx
@@ -5,7 +5,7 @@ import {filmProps} from "../film-screen/film-screen.prop";
 import {Link} from "react-router-dom";
 import User from "../user/user";
 import Logo from "../logo/logo";
-import CatalogMain from "../catalog/catalog-main";
+import Catalog from "../catalog/catalog";
 
 const MainScreen = ({films, poster, user}) => {
   return <React.Fragment>
@@ -74,7 +74,43 @@ const MainScreen = ({films, poster, user}) => {
       </div>
     </section>
     <div className="page-content">
-      <CatalogMain films={films} />
+      <Catalog films={films}>
+        <ul className="catalog__genres-list">
+        <li className="catalog__genres-item catalog__genres-item--active">
+          <a href="#" className="catalog__genres-link">All genres</a>
+        </li>
+        <li className="catalog__genres-item">
+          <a href="#" className="catalog__genres-link">Comedies</a>
+        </li>
+        <li className="catalog__genres-item">
+          <a href="#" className="catalog__genres-link">Crime</a>
+        </li>
+        <li className="catalog__genres-item">
+          <a href="#" className="catalog__genres-link">Documentary</a>
+        </li>
+        <li className="catalog__genres-item">
+          <a href="#" className="catalog__genres-link">Dramas</a>
+        </li>
+        <li className="catalog__genres-item">
+          <a href="#" className="catalog__genres-link">Horror</a>
+        </li>
+        <li className="catalog__genres-item">
+          <a href="#" className="catalog__genres-link">Kids &amp; Family</a>
+        </li>
+        <li className="catalog__genres-item">
+          <a href="#" className="catalog__genres-link">Romance</a>
+        </li>
+        <li className="catalog__genres-item">
+          <a href="#" className="catalog__genres-link">Sci-Fi</a>
+        </li>
+        <li className="catalog__genres-item">
+          <a href="#" className="catalog__genres-link">Thrillers</a>
+        </li>
+      </ul>
+        <div className="catalog__more">
+          <button className="catalog__button" type="button">Show more</button>
+        </div>
+      </ Catalog>
       <footer className="page-footer">
         <Logo />
         <div className="copyright">
