+++ src/utils/get-random-films.js
@@ -1,9 +1,4 @@
-export const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
+import {getRandomInteger} from "./get-random-integer";
 
 export const getRandomFilms = (films, count)=>{
   const lastIndex = films.length - 1;
